+++ src/main/java/com/bootcamp/blackbriar/auth/AuthenticationFilter.java
@@ -1,7 +1,7 @@
 package com.bootcamp.blackbriar.auth;
 
 import com.bootcamp.blackbriar.SpringApplicationContext;
-import com.bootcamp.blackbriar.business.user.UserService;
+import com.bootcamp.blackbriar.service.user.UserService;
 import com.bootcamp.blackbriar.model.user.UserLoginRequestModel;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.jsonwebtoken.Jwts;
