+++ src/main/java/com/bootcamp/blackbriar/controller/GroupController.java
@@ -1,6 +1,6 @@
-package com.bootcamp.blackbriar.api;
+package com.bootcamp.blackbriar.controller;
 
-import com.bootcamp.blackbriar.business.group.GroupService;
+import com.bootcamp.blackbriar.service.group.GroupService;
 import com.bootcamp.blackbriar.model.group.GroupEntity;
 import com.bootcamp.blackbriar.model.group.GroupResponse;
 import org.modelmapper.ModelMapper;
