+++ src/main/java/com/bootcamp/blackbriar/service/Utils.java
@@ -1,4 +1,4 @@
-package com.bootcamp.blackbriar.business;
+package com.bootcamp.blackbriar.service;
 
 import org.springframework.stereotype.Component;
