+++ src/main/java/com/bootcamp/blackbriar/service/group/GroupService.java
@@ -1,4 +1,4 @@
-package com.bootcamp.blackbriar.business.group;
+package com.bootcamp.blackbriar.service.group;
 
 import com.bootcamp.blackbriar.model.group.GroupEntity;
